Bug 1891303 - Update libwebp to v1.4.0 r=tnikkel
Differential Revision: https://phabricator.services.mozilla.com/D207359

UltraBlame original commit: 27b9c191fffa687683e6115a589a7859f473b5c1

+++ media/libwebp/src/dsp/lossless_enc_mips32.c
@@ -104,7 +104,7 @@ static float FastLog2Slow_MIPS32(uint32_t v) {
 
 
 
-static float ExtraCost_MIPS32(const uint32_t* const population, int length) {
+static uint32_t ExtraCost_MIPS32(const uint32_t* const population, int length) {
   int i, temp0, temp1;
   const uint32_t* pop = &population[4];
   const uint32_t* const LoopEnd = &population[length];
@@ -130,7 +130,7 @@ static float ExtraCost_MIPS32(const uint32_t* const population, int length) {
     : "memory", "hi", "lo"
   );
 
-  return (float)((int64_t)temp0 << 32 | temp1);
+  return ((int64_t)temp0 << 32 | temp1);
 }
 
 
@@ -149,8 +149,8 @@ static float ExtraCost_MIPS32(const uint32_t* const population, int length) {
 
 
 
-static float ExtraCostCombined_MIPS32(const uint32_t* const X,
-                                      const uint32_t* const Y, int length) {
+static uint32_t ExtraCostCombined_MIPS32(const uint32_t* const X,
+                                         const uint32_t* const Y, int length) {
   int i, temp0, temp1, temp2, temp3;
   const uint32_t* pX = &X[4];
   const uint32_t* pY = &Y[4];
@@ -183,7 +183,7 @@ static float ExtraCostCombined_MIPS32(const uint32_t* const X,
     : "memory", "hi", "lo"
   );
 
-  return (float)((int64_t)temp0 << 32 | temp1);
+  return ((int64_t)temp0 << 32 | temp1);
 }
 
 #define HUFFMAN_COST_PASS                                 \
